Fix test suite
The Material UI library doesn't seem to handle DOMless tests very well. It
requires a couple of attributes set in the child context of every component
(namely, "muiTheme") that breaks "TestUtils.renderIntoDocument" whenever one
is trying to render an inner component.

Hence the hack implemented in this commit: we added a "setMuiTheme" helper
that monkey-patches a given component so that the missing attributes are
properly set in the React class.