Improve PeopleAvailable tests
The Flux/reflux force us to separate the view-related code from the
store-related code. That enabled us to write unit tests for each module
separately. This commit contains a first draft of what we believe should be
tested in each module (and also how it should be tested).

We've upgraded the jasmine/karma-jasmine dependencies to make use of the
latest jasmine async-tests features.