Refactor Login views and Auth store
Use reflux to implement the Auth store, and split the Actions from the views.
The AuthStore tests are passable, but the auth view tests are horribly
incomplete. That's because we could not figure out how to properly test the
React router (including conditional rendering, redirecting and the
Authentication mixin). There's plenty of open bugs and half-baked solutions on
Github gists but they don't seem to work or test what we want.