Add redis and implement '/v1/consultants' API
Issue #70 require a couple of significant changes on the backend:
  1. a storage provider to save the "interviewer experience" of each
     consultant
  2. an API that allows us to query the list of consultants (optionally
     filtering by name & email) and also updates two attributes (the "p2
     interview experience level" and "p3 interview experience level")

We've decided to use Redis as a storage provider, so the Vagrant configuration
and the Gemfile have been updated to install and configure the addicional
dependency.

The API was implemented following the same patterns used for the calendar,
offices and roles APIs. The Consultant resource was heavily, though, because
we now have to provide an interface to save to/load from Redis, and also merge
the data coming from jigsaw with the extra attributes persisted locally. For
more details concerning the new API, check the file 'api/v1/consultants/
resource.rb'.

TODO: update the README and update the Vagrant configuration to seed the Redis
database when provisioning the system.